+++ PicoXLSX/LowLevel.cs
@@ -1,6 +1,6 @@
 ﻿/*
  * PicoXLSX is a small .NET library to generate XLSX (Microsoft Excel 2007 or newer) files in an easy and native way
- * Copyright Raphael Stoeckli © 2017
+ * Copyright Raphael Stoeckli © 2018
  * This library is licensed under the MIT License.
  * You find a copy of the license in project folder or on: http://opensource.org/licenses/MIT
  */
@@ -306,6 +306,7 @@ namespace PicoXLSX
             {
                 FileStream fs = new FileStream(this.workbook.Filename, FileMode.Create);
                 SaveAsStream(fs);
+                
             }
             catch (Exception e)
             {
@@ -385,6 +386,8 @@ namespace PicoXLSX
 
                     p.Flush();
                     p.Close();
+                    stream.Flush();
+                    stream.Close();
                 }
             }
             catch (Exception e)