+++ PicoXLSX/Style.cs
@@ -1,4 +1,11 @@
-﻿using System;
+﻿/*
+ * PicoXLSX is a small .NET library to generate XLSX (Microsoft Excel 2007 or newer) files in an easy and native way
+ * Copyright Raphael Stoeckli © 2018
+ * This library is licensed under the MIT License.
+ * You find a copy of the license in project folder or on: http://opensource.org/licenses/MIT
+ */
+
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;